initial commit
I'm finally putting this project in source control!  Arcadia is an old
project of mine... I see the oldest file creation date is July 24, 2009.
Before that, it was FRI, the Fantastic Rom Indexer.  I haven't worked on
Arcadia in over a year, but hopefully now I'll get to updating it!
http://davidmcgrath.com/arcadia